v0.3.2: modularized code
