+++ server/Makefile
@@ -1,6 +1,5 @@
 all:
 	@gcc -o server \
 		server.c \
-		read_peer_list.c \
-		get_address_info.c \
-		display_peer_info.c
+		peer_handling.c \
+		error_handle.c
