--- server/display_peer_info.h
@@ -1,8 +1,0 @@
-#ifndef DISPLAY_PEER_INFO_H
-#define DISPLAY_PEER_INFO_H
-
-#include "header.h"
-
-void display_peer_info(struct addrinfo *peers[PEER_NO]);
-
-#endif
