+++ server/error_handle.h
@@ -1,0 +1,10 @@
+#ifndef ERROR_HANDLE_H
+#define ERROR_HANDLE_H
+
+#include "header.h"
+
+void close_fd(int *fd, int size);
+
+void error_handle(char *context);
+
+#endif
