--- server/get_address_info.h
@@ -1,8 +1,0 @@
-#ifndef GET_ADDRESS_INFO_H 
-#define GET_ADDRESS_INFO_H
-
-#include "header.h"
-
-void get_address_info(struct addrinfo **peer, char *peer_ip, char *peer_port);
-
-#endif
