+++ server/header.h
@@ -12,9 +12,10 @@
 #include <arpa/inet.h>
 
 
-#define BUF_SIZE 100
-#define PEER_NO 10
-#define IP_WIDTH 16
-#define PORT_WIDTH 5
+#define BUF_SIZE 100            /* size of string buffer */
+#define PEER_NO 10              /* max number of peers */
+#define IP_WIDTH 16             /* width of ip addr string */
+#define PORT_WIDTH 5            /* width of port string */
+#define FD_NO PEER_NO+1         /* max number of file descriptors */
 
 #endif
